# licne-finansije/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Veb aplikacija za pregled licnih finansija",
        "description": "API za upravljanje ličnim finansijama (autentifikacija, transakcije, kategorije, budžeti, finansijski ciljevi, podsetnici, dokumenta i admin statistike)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
...
        "/api/finansijski-ciljevi": {
            "get": {
                "tags": [
                    "Finansijski Ciljevi"
                ],
                "summary": "Pregled svih finansijskih ciljeva",
                "description": "Vraća listu svih finansijskih ciljeva kreiranih u sistemu.",
                "operationId": "getFinansijskiCiljeviIndex",
                "responses": {
                    "200": {
                        "description": "Uspešno učitana lista finansijskih ciljeva",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "naziv": {
                                                "type": "string",
                                                "example": "Kupovina automobila"
                                            },
                                            "ciljaniIznos": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 5000
                                            },
                                            "trenutniIznos": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 1200.5
                                            },
                                            "rok": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2026-12-31"
                                            },
                                            "idKorisnik": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2026-06-18T20:00:00.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Finansijski Ciljevi"
                ],
                "summary": "Kreiranje novog finansijskog cilja",
                "description": "Omogućava postavljanje novog finansijskog cilja (npr. štednja za auto, letovanje) sa ciljanim iznosom, trenutnim stanjem i rokom.",
                "operationId": "createFinansijskiCilj",
                "requestBody": {
                    "description": "Podaci za kreiranje novog finansijskog cilja",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "idKorisnik",
                                    "naziv",
                                    "ciljni_iznos",
                                    "trenutni_iznos",
                                    "rok"
                                ],
                                "properties": {
                                    "idKorisnik": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "naziv": {
                                        "type": "string",
                                        "example": "Putovanje na Siciliju"
                                    },
                                    "ciljni_iznos": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1500,
                                        "minimum": 0
                                    },
                                    "trenutni_iznos": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 200,
                                        "minimum": 0
                                    },
                                    "rok": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-25"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Finansijski cilj uspešno kreiran",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "idKorisnik": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "naziv": {
                                            "type": "string",
                                            "example": "Putovanje na Siciliju"
                                        },
                                        "ciljni_iznos": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 1500
                                        },
                                        "trenutni_iznos": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 200
                                        },
                                        "rok": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2026-07-25"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2026-06-18T20:05:00.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validacija nije prošla",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validacija nije prosla"
                                        },
                                        "errors": {
                                            "properties": {
                                                "ciljni_iznos": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The ciljni iznos field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/finansijski-ciljevi/korisnik/{id}": {
            "get": {
                "tags": [
                    "Finansijski Ciljevi"
                ],
                "summary": "Pregled finansijskih ciljeva za određenog korisnika",
                "description": "Vraća listu svih finansijskih ciljeva koji pripadaju prosleđenom ID-ju korisnika.",
                "operationId": "getFinansijskiCiljeviByUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID korisnika čije finansijske ciljeve tražimo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno učitani ciljevi korisnika",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "naziv": {
                                                "type": "string",
                                                "example": "Letovanje"
                                            },
                                            "ciljaniIznos": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 1500
                                            },
                                            "trenutniIznos": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 600
                                            },
                                            "rok": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2026-07-20"
                                            },
                                            "idKorisnik": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2026-06-18T20:00:00.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    },
                    "404": {
                        "description": "Korisnik nije pronađen"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}